refactor(desktop): move createBrowserTab (the heavyweight) to tab-state (BP-P5b-7)
The single biggest tab function — 290 lines of BrowserView construction +
event-handler wiring (before-input-event, dom-ready, did-stop-loading,
context-menu, did-fail-load, etc.). All migrate.

New deps to support it:
- shouldAllowBrowserPopupWindow (utils — popup window heuristic)
- maybePromptBrowserInterrupt (P6 — user-vs-agent control gate)
- isAbortedBrowserLoadFailure (utils — did-fail-load filter)
- isProgrammaticBrowserInput (utils + main's WeakMap — input event gate)
- preloadDir (browserPopupPreload.cjs lookup for spawned popups)
- newTabTitle (NEW_TAB_TITLE constant)

Extended BrowserTabSpaceState structural type with lifecycleState (the
function flips it to 'active'). Extended BrowserWorkspaceState with
session + browserIdentity (used to construct the BrowserView's
webPreferences.session and seed its userAgent).

closeBrowserTab's last-tab-fallback createBrowserTab call is now an
internal direct call rather than a dep callback (the cyclic dep
collapses since both ends are local).

main.ts: 22,116 → 21,836 (-280 lines)
